flame & onphone base yolov8

See merge request r_d_center/department-of-ai/AlgorithmZoo!217

+++ cpp/onphone/box_info_impl.cpp
@@ -33,7 +33,7 @@ namespace glasssix::onphone
 		return internal_.y2;
 	}
 
-	int box_info_impl::category() const
+	float box_info_impl::category() const
 	{
 		return internal_.category;
 	}
